requirements.txt
requests==2.32.0
psycopg2-binary==2.9.10
python-dateutil==2.8.2
tqdm==4.66.1
